// src/models/dashboard/jobs/getCreatedJobs.model.tsx
export interface IGetCreatedJobs {
  title: string;
  institutionName: string;
  yearsOfExperience: string;
  pay: number;
  payInterval: string;
  location: string;
  jobDescription: any;
  isPublished: boolean;
  expiryDate: string;
  id: string;
  jobStartDate: string;
  jobEndDate: string;
  jobCreatorId: string;
}